Simplified the repository layout and user documentation
