+++ docs/README.md
@@ -1,0 +1,11 @@
+# Tsugumori guides
+
+- [Customize the desktop](customize.md)
+- [Controls and keybindings](controls.md)
+- [Send and receive files with Quickshare](quickshare.md)
+- [Common problems and recovery](help.md)
+- [Tested software versions](versions.md)
+- [v0.1.2 release notes](releases/v0.1.2.md)
+
+For installation, screenshots, and the project overview, return to the
+[main README](../README.md).
