+++ docs/controls.md
@@ -1,0 +1,37 @@
+# Controls and keybindings
+
+## Desktop shortcuts
+
+| Shortcut | Action |
+|---|---|
+| `SUPER` | Open the application menu |
+| `SUPER + Tab` | Open the Control Center |
+| `SUPER + L` | Lock the session |
+| `SUPER + T` | Open Kitty |
+| `SUPER + Return` | Show or hide the player |
+| `SUPER + R` | Restart only the desktop shell |
+| `SUPER + P` | Open the wallpaper picker |
+| `SUPER + Q` | Close the active window |
+| `SUPER + F` | Toggle fullscreen |
+| `ALT + Tab` | Cycle through windows |
+| `ALT + 1`, `2`, `3`... | Switch workspace |
+| `Print` | Take a screenshot |
+| `ALT + SHIFT + S` | Select an area for a screenshot |
+
+## Control Center
+
+Open the Control Center with `SUPER + Tab`. Its four areas provide:
+
+- **Connection** — Wi-Fi and Bluetooth
+- **Audio** — output selection, volume, and mute
+- **Quickshare** — send or receive files using HTTP and QR codes
+- **Notifications** — history, actions, Do Not Disturb, and Clear All
+
+See the [Quickshare guide](quickshare.md) for transfer steps, limits, and
+network-safety information.
+
+Use `W`, `A`, `S`, `D` or the arrow keys to move. Press `Enter` or `Space` to
+open the selected item, and press `Esc` to go back or close the menu.
+
+From the center, press a direction to open that side. Inside a panel, use the
+same keys to move through its controls.
