+++ maintenance/wave-assets/extract_last_frame.py
